fix warning & change array allocation method
fixed the compile warnings e.g. "include stdlib.h for malloc", "correct the type convert in front of malloc", "function return value mu_and_sig"

most importantly, change the way of allocating memory of arrays from array[m][n] (stack) to malloc.

+++ 2d_array.h
@@ -1,2 +1,3 @@
+#include "lib.h"
 int** int_2d_array_allocate(int r, int c);
-double** double_2d_array_allocate(double r, double c);
+double** double_2d_array_allocate(int r, int c);
